+++ app/tests/as_email/test_manage_provider_domains_command.py
@@ -19,7 +19,7 @@ from pytest_mock import MockerFixture
 # Project imports
 #
 from as_email.models import EmailAccount, Provider, Server
-from as_email.tests.factories import DummyProviderBackend
+from tests.as_email.factories import DummyProviderBackend
 
 pytestmark = pytest.mark.django_db